Better Command System
Command names are now stored in commands.json. This should make it easier to add new command names/aliases for commands plus if the command identifier needs to be changed it will be easier to do so. It should also make it easier to implement a better permission system.

All prior commands have been implemented into the new system.